public/python/requirements.txt
absl-py==2.1.0
arrow==1.3.0
asgiref==3.8.1
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
beautifulsoup4==4.12.3
binaryornot==0.4.4
bleach==6.1.0
blinker==1.6.2
cachetools==5.3.3
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.0
contourpy==1.2.0
cookiecutter==2.6.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
distlib==0.3.8
Django==5.0.4
django-browser-reload==1.12.1
django-tailwind==3.8.0
executing==2.0.1
filelock==3.13.1
Flask==2.3.3
flatbuffers==24.3.7
fonttools==4.47.0
gast==0.5.4
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
google_images_download==2.8.0
graphviz==0.20.1
grpcio==1.62.0
h11==0.14.0
h5py==3.10.0
idna==3.6
imageio==2.34.1
ipykernel==6.27.1
ipython==8.18.1
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
joblib==1.3.2
jupyter_client==8.6.0
jupyter_core==5.5.1
kaggle==1.6.12
keras==2.15.0
kiwisolver==1.4.5
lazy_loader==0.4
libclang==16.0.6
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.2.0
nest-asyncio==1.5.8
networkx==3.3
numpy==1.26.2
oauthlib==3.2.2
opencv-python==4.9.0.80
opendatasets==0.1.22
opt-einsum==3.3.0
outcome==1.3.0.post0
packaging==23.2
pandas==2.1.4
parso==0.8.3
patsy==0.5.6
Pillow==10.1.0
pipenv==2023.11.17
platformdirs==4.1.0
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.7
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.22
pydotplus==2.0.2
Pygments==2.17.2
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.4
pytz==2023.3.post1
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.2
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.7.1
rsa==4.9
scikit-image==0.23.2
scikit-learn==1.3.2
scipy==1.11.4
selenium==4.20.0
six==1.16.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
sqlparse==0.4.4
stack-data==0.6.3
statsmodels==0.14.1
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-estimator==2.15.0
tensorflow-intel==2.15.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
text-unidecode==1.3
threadpoolctl==3.2.0
tifffile==2024.5.10
tornado==6.4
tqdm==4.66.4
traitlets==5.14.0
trio==0.25.0
trio-websocket==0.11.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2023.3
urllib3==2.2.1
virtualenv==20.25.0
wcwidth==0.2.12
webencodings==0.5.1
Werkzeug==2.3.7
wrapt==1.14.1
wsproto==1.2.0
